// src/types/DeveloperInformation.ts
type DeveloperInformation = {
    login: string
    avatar_url: string
    html_url: string
    name: string
    company: string
    blog: string
    location: string
    bio: string
    twitter_username: string
    public_repos: number
    followers: number
    following: number
    created_at: string
}

export default DeveloperInformation